requirements.txt
gppt
pixivpy
stdiomask
